scx_layered: Move layer core growth logic to separate module
Move layer core growth logic to separate module for further refactoring.
